% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-core.R
\name{.is_gaussian}
\alias{.is_gaussian}
\title{Is a family Gaussian (NULL is treated as Gaussian)?}
\usage{
.is_gaussian(family)
}
\description{
Is a family Gaussian (NULL is treated as Gaussian)?
}
\keyword{internal}
