% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-mbco.R
\name{.mbco_gll}
\alias{.mbco_gll}
\title{Gaussian profile log-likelihood given SSE, variance V, and n rows.}
\usage{
.mbco_gll(sse, V, n)
}
\description{
Gaussian profile log-likelihood given SSE, variance V, and n rows.
}
\keyword{internal}
