% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-mbco.R
\name{.ols_sse}
\alias{.ols_sse}
\title{Residual sum of squares of an OLS fit of y on design matrix D.}
\usage{
.ols_sse(D, y)
}
\description{
Residual sum of squares of an OLS fit of y on design matrix D.
}
\keyword{internal}
