% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-core.R
\name{.pmed_compute}
\alias{.pmed_compute}
\title{Core P_med Computation Dispatcher}
\usage{
.pmed_compute(extract, x_ref, x_value, method, n_boot, ci_level, seed, ...)
}
\description{
Core P_med Computation Dispatcher
}
\keyword{internal}
