% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-parallel.R
\name{.pmed_compute_parallel}
\alias{.pmed_compute_parallel}
\title{Dispatch parallel (joint) P_med by inference method.}
\usage{
.pmed_compute_parallel(
  extract,
  x_ref,
  x_value,
  method,
  n_boot,
  ci_level,
  seed,
  ...
)
}
\description{
Dispatch parallel (joint) P_med by inference method.
}
\keyword{internal}
