% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-bootstrap.R
\name{.pmed_nonparametric_boot}
\alias{.pmed_nonparametric_boot}
\title{Nonparametric Bootstrap for P_med}
\usage{
.pmed_nonparametric_boot(extract, x_ref, x_value, n_boot, ci_level, seed, ...)
}
\description{
Nonparametric Bootstrap for P_med
}
\keyword{internal}
