% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-parallel.R
\name{.pmed_parallel_plugin}
\alias{.pmed_parallel_plugin}
\title{Plugin (point estimate only) for parallel joint P_med.}
\usage{
.pmed_parallel_plugin(extract, x_ref, x_value, ...)
}
\description{
Plugin (point estimate only) for parallel joint P_med.
}
\keyword{internal}
