% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-bootstrap.R
\name{.pmed_parametric_boot}
\alias{.pmed_parametric_boot}
\title{Parametric Bootstrap for P_med}
\usage{
.pmed_parametric_boot(extract, x_ref, x_value, n_boot, ci_level, seed, ...)
}
\description{
Parametric Bootstrap for P_med
}
\keyword{internal}
