% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-core.R
\name{.pmed_plugin}
\alias{.pmed_plugin}
\title{Plugin Estimator (Point Estimate Only)}
\usage{
.pmed_plugin(extract, x_ref, x_value, n_sim = 10000, ...)
}
\description{
Plugin Estimator (Point Estimate Only)
}
\keyword{internal}
