% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-parallel-mbco.R
\name{.qr_sse}
\alias{.qr_sse}
\title{Residual sum of squares against a precomputed QR decomposition.}
\usage{
.qr_sse(qr_obj, y)
}
\description{
Residual sum of squares against a precomputed QR decomposition.
}
\keyword{internal}
