% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{data}
\name{lm_class}
\alias{lm_class}
\title{S3 Class Wrappers for S7}
\format{
An object of class \code{S7_S3_class} of length 3.
}
\usage{
lm_class
}
\description{
S7 wrappers for S3 classes used in method dispatch
}
\keyword{internal}
