% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-core.R
\name{.pmed_core_simple}
\alias{.pmed_core_simple}
\title{Simple P_med Core Computation}
\usage{
.pmed_core_simple(
  a,
  b,
  c_prime,
  x_ref,
  x_value,
  sigma_y = 1,
  sigma_m = 1,
  n_sim = 10000
)
}
\description{
Simple P_med Core Computation
}
\keyword{internal}
