% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{extract_mediation}
\alias{extract_mediation}
\title{Extract Mediation Structure}
\usage{
extract_mediation(object, ...)
}
\arguments{
\item{object}{Fitted model object}

\item{...}{Additional arguments passed to methods}
}
\value{
MediationExtract object
}
\description{
Generic function to extract mediation structure from fitted models.
See method documentation for specific arguments.
}
