% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{pmed}
\alias{pmed}
\title{Compute P_med}
\usage{
pmed(object, ...)
}
\arguments{
\item{object}{Either a formula, MediationExtract, or fitted model}

\item{...}{Additional arguments passed to methods}
}
\value{
PmedResult object
}
\description{
Generic function to compute P_med from various inputs.
}
