% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-imports.R, R/probmed-package.R
\docType{package}
\name{probmed-package}
\alias{probmed-package}
\alias{probmed}
\title{probmed: Probabilistic Effect Sizes for Causal Mediation Analysis}
\description{
Compute P_med, a scale-free probabilistic effect size for causal mediation analysis.

Compute P_med, a scale-free probabilistic effect size for causal mediation
analysis. P_med = P(Y_\{x*,M_x\} > Y_\{x,M_x\}) represents the probability that
the outcome under treatment with the mediator at its control level exceeds
the outcome under control.

Compute P_med, a scale-free probabilistic effect size for causal mediation analysis.
}
\author{
\strong{Maintainer}: Davood Tofighi \email{dtofighi@gmail.com} (\href{https://orcid.org/0000-0001-8523-7776}{ORCID})

}
\keyword{internal}
